Fix prevent overwriting state set while suspended
Due to some unknown race condition, if the application is launched from
the launcher while the screen is locked, the app might set the camera
state after the application state become Suspended. In this case, we
don't want to overwrite the state with the state before suspended when
the app become active again.

The old code assumes that in this case, the app will never become active
before suspended. However, the recent change in QtUbuntu [1] breaks this
assumption. This commit adds another condition: it will not overwrite
the state when the app is active if it's set by the app. This requires
moving the application state handling to CameraControl in order to
distinguish between setting state from the app and from itself.

[1] https://github.com/ubports/qtubuntu/pull/14

Fixes: https://github.com/ubports/ubuntu-touch/issues/1403